Move DNS trace latency calculation to BPF
Previously, DNS trace latency was calculated in userspace
by inserting and retrieving events into an LRU hashmap.
This implicitly assumed that events were ordered ascending by timestamp,
but BPF perf ring buffers do not guarantee chronological order
(see https://kinvolk.io/blog/2018/02/timing-issues-when-using-bpf-with-virtual-cpus/)

This resulted in missing latency fields in the DNS trace events
and caused the TestTraceDns to fail randomly, as reported in
https://github.com/inspektor-gadget/inspektor-gadget/issues/1440

To address this problem, move the latency calculation from userspace
to the BPF program, with query timestamps stored in a BPF map.
To ensure that queries without a response are removed from the map,
run a background garbage collection goroutine to remove timestamps
older than a threshold.

Signed-off-by: Will Daly <widaly@microsoft.com>